no need to delete if there is no receipt

+++ finance_site/finance/views/UpdateTransaction.py
@@ -57,7 +57,7 @@ class UpdateTransaction(View):
         transaction.store = request.POST['store']
         transaction.note = request.POST['note']
         if file_name is not None:
-            if (file_name != transaction.receipt):
+            if transaction.receipt is not None and (file_name != transaction.receipt):
                 fs.delete(transaction.receipt.name)
             transaction.receipt = file_name
         transaction.categories = Transaction.objects.get(id=request.POST['categories'])
